+++ menu.py
@@ -1,13 +1,22 @@
 import streamlit as st
-if "current_page" not in st.session_state:
-    st.session_state.current_page = "chat"
 def menu()-> None:
     try:
         st.set_page_config(layout = "wide")
     except:
         pass
     st.sidebar.image("images/logo.png")
-    st.sidebar.title("Welcome to ClerkGPT!")
+    st.sidebar.header(f"Welcome to ClerkGPT, {st.user.name}!")
+    if st.sidebar.button("Log out"):
+        st.logout()
+        st.session_state["authenticated"] = False
+        st.session_state["made_attempt"] = False
+        # --- Optional: Reset conversation ---
+    if 'current_page' not in st.session_state:
+        st.session_state.current_page = 'chat'
+    if st.session_state.current_page == 'chat':
+        if st.sidebar.button("Reset Conversation"):
+            st.session_state.messages = []
+            st.session_state.memory.clear()
     st.sidebar.page_link("pages/about.py", label="About", icon="ℹ️")
     st.sidebar.page_link("pages/chat.py", label="Chat", icon="🤖")
     st.sidebar.page_link("pages/faq.py", label="FAQ", icon="❓")