+++ changelog.md
@@ -2,15 +2,26 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2025-06-30
+### Added
+- Google authentication and Astra DB used for authentication
+- Added tracking queries
+
+### Changed
+- Moved reset button to chat page only
+- Authentication based on database query and Google Auth
+- Changed number of references from 4 to 10
+
 
 ## [1.0.0] - 2025-06-29
 ### Added
 - Initial release of ClerkGPT chat app.
-- User authentication via secrets.
+- User authentication via secrets and Google Authentication.
 - Conversational memory with LangChain.
 - Reference rendering for source documents.
 - Reset conversation button
 - Added caching of vector store
+- Added 52nd GA documents
 
 ### Changed
 - Changed how references was created